requirements.txt
opencv-python-headless
tcolorpy
paho-mqtt
numpy
flask
